First working version of the ondemand implementation
Implemented an ondemand dynamic version of task assignments for
automatic load balancing

+++ src/ondemand/worker.h
@@ -13,7 +13,7 @@ class Worker {
         void compute_mandelbrot(int);
         /* will contain both lower and upper bound of the image's partition to compute */
         unsigned from_to[2];
-        std::vector<double> result;
+        std::vector<int> result;
         unsigned width, height;
         unsigned master_rank;
         int rank;
